requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
requests==2.31.0
pandas==2.1.0
numpy==1.25.2
transformers==4.33.1
python-multipart==0.0.6
aiohttp==3.8.5
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.0
pydantic-settings==2.0.3
PyPDF2==3.0.1
pdfminer.six==20221105
python-magic-bin==0.4.14